app/requirements.txt
fastapi
uvicorn
graphene
graphene-mongo
mongoengine
pydantic
strawberry-graphql